Adds map click to creatingTrail integration test
- Refactors the code for creating trail components so that the state
is passed down from the parent component rather than other way around.
